move from src directory to remove stories from bundle
