+++ stories/accordion/RawAccordion.stories.tsx
@@ -1,28 +1,27 @@
+import React from "react";
 import type { Meta, StoryObj } from "@storybook/react";
-
-import { Alert } from "react-native";
-import { H2 } from "../../components";
+import { AccordionItem, Caption } from "../../src/components";
 
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 const meta = {
-  title: "IO-App Design System/Atoms/Typography/H2",
-  component: H2,
+  title: "IO-App Design System/Atoms/Accordion/Accordion",
+  component: AccordionItem,
   parameters: {
     // Optional parameter to center the component in the Canvas. More info: https://storybook.js.org/docs/react/configure/story-layout
-    layout: "centered",
-    actions: { argTypesRegex: "^on.*" }
+    layout: "padded"
   },
   // This component will have an automatically generated Autodocs entry: https://storybook.js.org/docs/react/writing-docs/autodocs
   tags: ["autodocs"]
-} satisfies Meta<typeof H2>;
+} satisfies Meta<typeof AccordionItem>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
 
-export const Primary: Story = {
+// More on writing stories with args: https://storybook.js.org/docs/react/writing-stories/args
+export const BaseAccordion: Story = {
   args: {
-    accessibilityLabel: "Tap to trigger test alert",
-    onPress: () => Alert.alert("Alert", "Action triggered"),
-    children: "Hello World"
+    title: "Accordion Title",
+    id: 1,
+    body: <Caption>{"Accordion body"}</Caption>
   }
 };
