+++ stories/buttons/ButtonExtendedOutline.stories.ts
@@ -1,28 +1,29 @@
 import type { Meta, StoryObj } from "@storybook/react";
+import { action } from "@storybook/addon-actions";
 
-import { Alert } from "react-native";
-import { H5 } from "../../components";
+import { ButtonExtendedOutline } from "../../src/components";
 
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 const meta = {
-  title: "IO-App Design System/Atoms/Typography/H5",
-  component: H5,
+  title: "IO-App Design System/Atoms/Buttons/ButtonExtendedOutline",
+  component: ButtonExtendedOutline,
   parameters: {
     // Optional parameter to center the component in the Canvas. More info: https://storybook.js.org/docs/react/configure/story-layout
-    layout: "centered",
-    actions: { argTypesRegex: "^on.*" }
+    layout: "centered"
   },
   // This component will have an automatically generated Autodocs entry: https://storybook.js.org/docs/react/writing-docs/autodocs
   tags: ["autodocs"]
-} satisfies Meta<typeof H5>;
+} satisfies Meta<typeof ButtonExtendedOutline>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
 
+// More on writing stories with args: https://storybook.js.org/docs/react/writing-stories/args
 export const Primary: Story = {
   args: {
-    accessibilityLabel: "Tap to trigger test alert",
-    onPress: () => Alert.alert("Alert", "Action triggered"),
-    children: "Hello World"
+    label: "Primary button",
+    onPress: e => {
+      action("clicked")(e);
+    }
   }
 };
