+++ stories/buttons/ButtonOutline.stories.ts
@@ -1,19 +1,19 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import { action } from "@storybook/addon-actions";
 
-import { ButtonLink } from "../../components";
+import { ButtonOutline } from "../../src/components";
 
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 const meta = {
-  title: "IO-App Design System/Atoms/Buttons/ButtonLink",
-  component: ButtonLink,
+  title: "IO-App Design System/Atoms/Buttons/ButtonOutline",
+  component: ButtonOutline,
   parameters: {
     // Optional parameter to center the component in the Canvas. More info: https://storybook.js.org/docs/react/configure/story-layout
     layout: "centered"
   },
   // This component will have an automatically generated Autodocs entry: https://storybook.js.org/docs/react/writing-docs/autodocs
   tags: ["autodocs"]
-} satisfies Meta<typeof ButtonLink>;
+} satisfies Meta<typeof ButtonOutline>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
@@ -24,6 +24,26 @@ export const Primary: Story = {
     color: "primary",
     accessibilityLabel: "Tap to trigger test alert",
     label: "Primary button",
+    onPress: e => {
+      action("clicked")(e);
+    }
+  }
+};
+export const Danger: Story = {
+  args: {
+    color: "danger",
+    accessibilityLabel: "Tap to trigger test alert",
+    label: "Danger button",
+    onPress: e => {
+      action("clicked")(e);
+    }
+  }
+};
+export const Contrast: Story = {
+  args: {
+    color: "contrast",
+    accessibilityLabel: "Tap to trigger test alert",
+    label: "Contrast button",
     onPress: e => {
       action("clicked")(e);
     }
