+++ stories/buttons/IconButtonContained.stories.ts
@@ -1,19 +1,19 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import { action } from "@storybook/addon-actions";
 
-import { ButtonSolid } from "../../components";
+import { IconButtonContained } from "../../src/components";
 
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 const meta = {
-  title: "IO-App Design System/Atoms/Buttons/ButtonSolid",
-  component: ButtonSolid,
+  title: "IO-App Design System/Atoms/Buttons/IconButtonContained",
+  component: IconButtonContained,
   parameters: {
     // Optional parameter to center the component in the Canvas. More info: https://storybook.js.org/docs/react/configure/story-layout
     layout: "centered"
   },
   // This component will have an automatically generated Autodocs entry: https://storybook.js.org/docs/react/writing-docs/autodocs
   tags: ["autodocs"]
-} satisfies Meta<typeof ButtonSolid>;
+} satisfies Meta<typeof IconButtonContained>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
@@ -22,28 +22,8 @@ type Story = StoryObj<typeof meta>;
 export const Primary: Story = {
   args: {
     color: "primary",
+    icon: "add",
     accessibilityLabel: "Tap to trigger test alert",
-    label: "Primary button",
-    onPress: e => {
-      action("clicked")(e);
-    }
-  }
-};
-export const Danger: Story = {
-  args: {
-    color: "danger",
-    accessibilityLabel: "Tap to trigger test alert",
-    label: "Danger button",
-    onPress: e => {
-      action("clicked")(e);
-    }
-  }
-};
-export const Contrast: Story = {
-  args: {
-    color: "contrast",
-    accessibilityLabel: "Tap to trigger test alert",
-    label: "Contrast button",
     onPress: e => {
       action("clicked")(e);
     }
