+++ stories/buttons/IconButtonSolid.stories.ts
@@ -1,29 +1,42 @@
 import type { Meta, StoryObj } from "@storybook/react";
+import { action } from "@storybook/addon-actions";
 
-import { Alert } from "react-native";
-import { Chip } from "../../components";
+import { IconButtonSolid } from "../../src/components";
 
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 const meta = {
-  title: "IO-App Design System/Atoms/Typography/Chip",
-  component: Chip,
+  title: "IO-App Design System/Atoms/Buttons/IconButtonSolid",
+  component: IconButtonSolid,
   parameters: {
     // Optional parameter to center the component in the Canvas. More info: https://storybook.js.org/docs/react/configure/story-layout
-    layout: "centered",
-    actions: { argTypesRegex: "^on.*" }
+    layout: "centered"
   },
   // This component will have an automatically generated Autodocs entry: https://storybook.js.org/docs/react/writing-docs/autodocs
   tags: ["autodocs"]
-} satisfies Meta<typeof Chip>;
+} satisfies Meta<typeof IconButtonSolid>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
 
+// More on writing stories with args: https://storybook.js.org/docs/react/writing-stories/args
 export const Primary: Story = {
   args: {
-    color: "black",
+    color: "primary",
+    icon: "add",
     accessibilityLabel: "Tap to trigger test alert",
-    onPress: () => Alert.alert("Alert", "Action triggered"),
-    children: "Hello World"
+    onPress: e => {
+      action("clicked")(e);
+    }
+  }
+};
+
+export const Contrast: Story = {
+  args: {
+    color: "contrast",
+    icon: "add",
+    accessibilityLabel: "Tap to trigger test alert",
+    onPress: e => {
+      action("clicked")(e);
+    }
   }
 };
