+++ stories/tsconfig.json
@@ -1,0 +1,3 @@
+{
+  "extends": "../tsconfig"
+}
