+++ stories/typography/Caption.stories.ts
@@ -1,12 +1,12 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
 import { Alert } from "react-native";
-import { LabelSmall } from "../../components";
+import { Caption } from "../../src/components";
 
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 const meta = {
-  title: "IO-App Design System/Atoms/Typography/LabelSmall",
-  component: LabelSmall,
+  title: "IO-App Design System/Atoms/Typography/Caption",
+  component: Caption,
   parameters: {
     // Optional parameter to center the component in the Canvas. More info: https://storybook.js.org/docs/react/configure/story-layout
     layout: "centered",
@@ -14,7 +14,7 @@ const meta = {
   },
   // This component will have an automatically generated Autodocs entry: https://storybook.js.org/docs/react/writing-docs/autodocs
   tags: ["autodocs"]
-} satisfies Meta<typeof LabelSmall>;
+} satisfies Meta<typeof Caption>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
