+++ stories/typography/H2.stories.ts
@@ -1,12 +1,12 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
 import { Alert } from "react-native";
-import { ButtonText } from "../../components";
+import { H2 } from "../../src/components";
 
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 const meta = {
-  title: "IO-App Design System/Atoms/Typography/ButtonText",
-  component: ButtonText,
+  title: "IO-App Design System/Atoms/Typography/H2",
+  component: H2,
   parameters: {
     // Optional parameter to center the component in the Canvas. More info: https://storybook.js.org/docs/react/configure/story-layout
     layout: "centered",
@@ -14,14 +14,13 @@ const meta = {
   },
   // This component will have an automatically generated Autodocs entry: https://storybook.js.org/docs/react/writing-docs/autodocs
   tags: ["autodocs"]
-} satisfies Meta<typeof ButtonText>;
+} satisfies Meta<typeof H2>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
 
 export const Primary: Story = {
   args: {
-    color: "black",
     accessibilityLabel: "Tap to trigger test alert",
     onPress: () => Alert.alert("Alert", "Action triggered"),
     children: "Hello World"
