+++ stories/typography/H3.stories.ts
@@ -1,27 +1,28 @@
-import React from "react";
 import type { Meta, StoryObj } from "@storybook/react";
-import { AccordionItem, Caption } from "../../components";
+
+import { Alert } from "react-native";
+import { H3 } from "../../src/components";
 
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 const meta = {
-  title: "IO-App Design System/Atoms/Accordion/Accordion",
-  component: AccordionItem,
+  title: "IO-App Design System/Atoms/Typography/H3",
+  component: H3,
   parameters: {
     // Optional parameter to center the component in the Canvas. More info: https://storybook.js.org/docs/react/configure/story-layout
-    layout: "padded"
+    layout: "centered",
+    actions: { argTypesRegex: "^on.*" }
   },
   // This component will have an automatically generated Autodocs entry: https://storybook.js.org/docs/react/writing-docs/autodocs
   tags: ["autodocs"]
-} satisfies Meta<typeof AccordionItem>;
+} satisfies Meta<typeof H3>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
 
-// More on writing stories with args: https://storybook.js.org/docs/react/writing-stories/args
-export const BaseAccordion: Story = {
+export const Primary: Story = {
   args: {
-    title: "Accordion Title",
-    id: 1,
-    body: <Caption>{"Accordion body"}</Caption>
+    accessibilityLabel: "Tap to trigger test alert",
+    onPress: () => Alert.alert("Alert", "Action triggered"),
+    children: "Hello World"
   }
 };
