+++ stories/typography/H6.stories.ts
@@ -1,12 +1,12 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
 import { Alert } from "react-native";
-import { Caption } from "../../components";
+import { H6 } from "../../src/components";
 
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 const meta = {
-  title: "IO-App Design System/Atoms/Typography/Caption",
-  component: Caption,
+  title: "IO-App Design System/Atoms/Typography/H6",
+  component: H6,
   parameters: {
     // Optional parameter to center the component in the Canvas. More info: https://storybook.js.org/docs/react/configure/story-layout
     layout: "centered",
@@ -14,7 +14,7 @@ const meta = {
   },
   // This component will have an automatically generated Autodocs entry: https://storybook.js.org/docs/react/writing-docs/autodocs
   tags: ["autodocs"]
-} satisfies Meta<typeof Caption>;
+} satisfies Meta<typeof H6>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
