+++ stories/typography/Hero.stories.ts
@@ -1,29 +1,29 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import { action } from "@storybook/addon-actions";
 
-import { ButtonExtendedOutline } from "../../components";
+import { Alert } from "react-native";
+import { Hero } from "../../src/components";
 
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 const meta = {
-  title: "IO-App Design System/Atoms/Buttons/ButtonExtendedOutline",
-  component: ButtonExtendedOutline,
+  title: "IO-App Design System/Atoms/Typography/Hero",
+  component: Hero,
   parameters: {
     // Optional parameter to center the component in the Canvas. More info: https://storybook.js.org/docs/react/configure/story-layout
-    layout: "centered"
+    layout: "centered",
+    actions: { argTypesRegex: "^on.*" }
   },
   // This component will have an automatically generated Autodocs entry: https://storybook.js.org/docs/react/writing-docs/autodocs
   tags: ["autodocs"]
-} satisfies Meta<typeof ButtonExtendedOutline>;
+} satisfies Meta<typeof Hero>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
 
-// More on writing stories with args: https://storybook.js.org/docs/react/writing-stories/args
 export const Primary: Story = {
   args: {
-    label: "Primary button",
-    onPress: e => {
-      action("clicked")(e);
-    }
+    color: "black",
+    accessibilityLabel: "Tap to trigger test alert",
+    onPress: () => Alert.alert("Alert", "Action triggered"),
+    children: "Hello World"
   }
 };
