+++ stories/typography/Monospace.stories.ts
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
 import { Alert } from "react-native";
-import { BodyMonospace } from "../../components";
+import { BodyMonospace } from "../../src/components";
 
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 const meta = {
